backend/requirements.txt
# Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.20
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
gunicorn>=21.2.0
